requirements.txt
asgiref==3.7.2
distlib==0.3.8
dj-database-url==2.1.0
Django==5.0.2
django-cors-headers==4.3.1
django-heroku==0.3.1
djangorestframework==3.14.0
filelock==3.13.1
gunicorn==21.2.0
packaging==23.2
platformdirs==4.2.0
psycopg2==2.9.9
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
virtualenv==20.25.0
whitenoise==6.6.0
